--amend

+++ deps_adapter/key_management_adapter/impl/src/standard/crypto_hash_to_point.c
@@ -314,12 +314,12 @@ ERR:
 static int32_t EndianSwap(struct HksBlob *data)
 {
     uint32_t end = data->size - 1;
-    const int32_t start = 0;
+    const uint32_t start = 0;
 
     /* count the middle index of array */
-    int32_t cnt = (int32_t)data->size / 2; // 2 used to calculate half of the data size
+    uint32_t cnt = data->size / 2; // 2 used to calculate half of the data size
 
-    for (int32_t i = 0; i < cnt; i++) {
+    for (uint32_t i = 0; i < cnt; i++) {
         uint8_t tmp;
         tmp = data->data[start + i];
         data->data[start + i] = data->data[end - i];
